requirements.txt
torch
torchvision
opencv-python
tqdm
flask
flask-cors
optuna
matplotlib
numpy
pillow
axios
kaggle
